% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{ante_conse}
\alias{ante_conse}
\title{Determine model antecedents and consequents}
\usage{
ante_conse(permute_df, model_index)
}
\arguments{
\item{permute_df}{permutation dataframe}

\item{model_index}{model index represented as a numeric vector}
}
\value{
list containing original model index, and the model's direct antecedents, consequents, and antecedents, each their own list, sorted by variable name
}
\description{
Given a permutation dataframe and model indices to generate a specific model, loops through all variables to determine their antecedents, consequents, and direct antecedents
}
