% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url.R
\name{api_url}
\alias{api_url}
\title{Get Gemini's API URL}
\usage{
api_url(
  gemini_api_key,
  endpoint = "generateContent",
  model = Sys.getenv("GEMINI_MODEL", "gemini-pro")
)
}
\arguments{
\item{gemini_api_key}{Gemini's API key.}

\item{endpoint}{Gemini's endpoint to target.}

\item{model}{Gemini's model to use.}
}
\description{
Get Gemini's API URL
}
