% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_gemini.R
\name{ask_gemini}
\alias{ask_gemini}
\title{Ask Gemini}
\usage{
ask_gemini(
  question,
  session_id = "1",
  gemini_api_key = Sys.getenv("GEMINI_API_KEY")
)
}
\arguments{
\item{question}{The question to ask Gemini.}

\item{session_id}{The ID of the session to be used.}

\item{gemini_api_key}{Gemini's API key.}
}
\value{
A character value with the response generated by Gemini.
}
\description{
Note: See also `reset_chat_session`.
}
\examples{
\dontrun{
cat(ask_gemini("What do you think about R language?"))
}

}
