% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{code.title}
\alias{code.title}
\title{A code.title function}
\usage{
code.title(what)
}
\arguments{
\item{what}{content you want to put in the title}
}
\description{
this function allows you to create title for codes with date
}

