% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.sets.R
\name{compare.sets}
\alias{compare.sets}
\title{A compare sets function}
\usage{
compare.sets(x, y)
}
\arguments{
\item{x}{the first set to be compared}

\item{y}{the second set to be compared}
}
\description{
This function allows you to compare two sets, get the number of unique
and common elements in each sets
}
\examples{
()
compare.sets(c(1,1,2,3,4), c(2,2,3,4,5))
}
\keyword{comparison}
\keyword{sets}

