% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Compound pipe}
\arguments{
\item{lhs, rhs}{An object modified in place, and a function to apply to it}
}
\description{
Compound pipe
}
\examples{
a <- 1:5
a \%<>\% sum() # a is modified in place

}
