% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condition_occurrence_data.R
\name{condition_occurrence_data}
\alias{condition_occurrence_data}
\title{Condition_occurrence_data}
\usage{
condition_occurrence_data()
}
\description{
This function extract data from condition occurrence table
}
\keyword{gemini}
