% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{counts}
\alias{counts}
\title{Big Papi counts matrix}
\format{An object of class \code{matrix} with 9216 rows and 16 columns.}
\usage{
counts
}
\description{
The counts matrix from the Big Papi SynLet library.
}
\keyword{data}
