% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{delavaanify}
\alias{delavaanify}
\title{"De"-lavaanify a parameter dataframe}
\usage{
delavaanify(param_df)
}
\arguments{
\item{param_df}{data frame detailing specified free parameters (i.e., include lhs, op, and rhs columns)}
}
\description{
given a param df, reverses the lavaanify function (see \code{\link[lavaan:lavaanify]{lavaan::lavaanify()}}), transforming back into a string
}
