% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{determine_LBR}
\alias{determine_LBR}
\title{Determine limited block-recursiveness}
\usage{
determine_LBR(model, row, model_info)
}
\arguments{
\item{model}{dataframe specifying overall model (i.e., all estimated parameters)}

\item{row}{dataframe of 1 row length specifying focal block under investigation}

\item{model_info}{model info indicating information regarding variable antecedents and consequents}
}
\value{
If focal block fails to pass limited block-recursiveness, returns FALSE. If focal block passes limited block-recursiveness and the generated preceding block is just-identified, outputs model names of all generated names from \code{\link[=pbl_jid_gen]{pbl_jid_gen()}}. Otherwise, returns TRUE.
}
\description{
Given a model, a specific row in the model, and that model's information (i.e., regarding antecedents and consequents), determines whether the chosen focal block (defined by the specific row) obeys limited block-recursiveness
}
