% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.median_normalize}
\alias{.median_normalize}
\title{Median normalization}
\usage{
.median_normalize(counts, CONSTANT = 32)
}
\arguments{
\item{counts}{a counts matrix derived from Input}

\item{CONSTANT}{a pseudo-count to use (default = 32)}
}
\value{
median-normalized counts object
}
\description{
Median normalization
}
\examples{
\dontrun{
.median_normalize(rnorm(n = 1000))
}
}
