% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_condition_occurrence.R
\name{draw_condition_occurrence}
\alias{draw_condition_occurrence}
\title{Draw condition occurrence}
\usage{
draw_condition_occurrence()
}
\description{
This function for draw graph from condition occurrence RDS data
}
\keyword{gemini}
