% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_drug_exposure.R
\name{draw_drug_exposure}
\alias{draw_drug_exposure}
\title{Draw drug exposure}
\usage{
draw_drug_exposure()
}
\description{
This function for draw graph from drug exposure RDS data
}
\keyword{gemini}
