% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_gender_bar}
\alias{draw_gender_bar}
\title{Draw bar plot about CDM person gender with plotly}
\usage{
draw_gender_bar(dbName, rds)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}
}
\value{
Return bar plot.
}
\description{
\code(draw_gender_bar) Draw bar plot about gender
}
\details{
\code(draw_gender_bar) Draw bar plot about gender
}
\examples{
\dontrun{
draw_gender_bar(dbName = 'MYCDM', rds = loaded_rds)
}
}
