% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_histogram}
\alias{draw_histogram}
\title{Draw bar plot about CDM data}
\usage{
draw_histogram(dbName, rds, rdsConcept)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}

\item{rdsConcept}{rds}
}
\value{
Return bar plot.
}
\description{
\code(draw_histogram) Draw bar plot
}
\details{
\code(draw_histogram) Draw bar plot
}
\examples{
\dontrun{
draw_histogram(dbName = 'MYCDM', rds = loaded_rds, rdsConcept = 'persontbl_location')
}
}
