% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_line_graph}
\alias{draw_line_graph}
\title{Draw line graph about CDM data}
\usage{
draw_line_graph(dbName, rds, rdsConcept)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}

\item{rdsConcept}{rds}
}
\value{
Return line graph
}
\description{
\code(draw_line_graph) Draw line graph
}
\details{
\code(draw_line_graph) Draw line graph
}
\examples{
\dontrun{
draw_line_graph(dbName = 'MYCDM', rds = loaded_rds, rdsConcept = 'conditiontbl_start')
}
}
