% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_meta_pie}
\alias{draw_meta_pie}
\title{Draw pie plot about CDM person, record count with plotly library}
\usage{
draw_meta_pie(dbName, rds, tableName, type)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}

\item{tableName}{CDM database table name}

\item{type}{"person" or "record"}
}
\value{
Return pie plot.
}
\description{
\code(draw_meta_pie) Draw pie plot about table person or record
}
\details{
\code(draw_meta_pie) Draw pie plot about table person or record
}
\examples{
\dontrun{
draw_meta_pie(dbName = 'MYCDM', rds = loaded_rds, tableName = 'person', type = 'person')
}
}
