% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_null_bar}
\alias{draw_null_bar}
\title{Draw bar plot with null ratio about CDM data}
\usage{
draw_null_bar(dbName, rds, rdsConcept)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}

\item{rdsConcept}{rds}
}
\value{
Return two bar plot.
}
\description{
\code(draw_null_bar) Draw bar plot with null ratio
}
\details{
\code(draw_null_bar) Draw bar plot with null ratio
}
\examples{
\dontrun{
draw_null_bar(dbName = 'MYCDM', rds = loaded_rds, rdsConcept = 'persontbl_location')
}
}
