% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_pie}
\alias{draw_pie}
\title{Draw pie plot about CDM data with plotly}
\usage{
draw_pie(dbName, rds, rdsConcept, title)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}

\item{rdsConcept}{rds}
}
\value{
Return pie plot.
}
\description{
\code(draw_pie) Draw bar plot about gender
}
\details{
\code(draw_pie) Draw bar plot about gender
}
\examples{
\dontrun{
draw_pie(dbName = 'MYCDM', rds = loaded_rds, rdsConcept = 'visittbl_visit_concept')
}
}
