% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_total_person}
\alias{draw_total_person}
\title{Draw bar plot about CDM people count with plotly library}
\usage{
draw_total_person(dbName, rds)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}
}
\value{
Return bar plot.
}
\description{
\code{draw_total_person} Draw bar plot about person count
}
\details{
\code{draw_total_person} Draw bar plot about person count
}
\examples{
\dontrun{
draw_total_person(dbName = 'MYCDM', rds = loaded_rds)
}
}
