% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_total_record}
\alias{draw_total_record}
\title{Draw bar plot about CDM record count with plotly library}
\usage{
draw_total_record(dbName, rds)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}
}
\value{
Return bar plot.
}
\description{
\code(draw_total_record) Draw bar plot about record count
}
\details{
\code(draw_total_record) Draw bar plot about record count
}
\examples{
\dontrun{
draw_total_record(dbName = 'MYCDM', rds = loaded_rds)
}
}
