% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_function.R
\name{draw_visit_freq_graph}
\alias{draw_visit_freq_graph}
\title{Draw bar plot about CDM person gender with plotly}
\usage{
draw_visit_freq_graph(dbName, rds, type)
}
\arguments{
\item{dbName}{CDM database name}

\item{rds}{loaded CDM data}

\item{type}{'min' or 'max'}
}
\value{
Return bar plot.
}
\description{
\code(draw_visit_freq_graph) Draw bar plot about gender
}
\details{
\code(draw_visit_freq_graph) Draw bar plot about gender
}
\examples{
\dontrun{
draw_visit_freq_graph(dbName = 'MYCDM', rds = loaded_rds, type = 'min')
}
}
