% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{equiv_gen}
\alias{equiv_gen}
\title{Generate Equivalent Models}
\usage{
equiv_gen(lav_spec)
}
\arguments{
\item{lav_spec}{\emph{lavaan} model specification}
}
\value{
list containing a) a permutation dataframe, b) a list of equivalent models, and c) names of invalid models found during the generation process. For more details, see (REDACTED, in prep.)
}
\description{
\code{equiv_gen()} takes in a \emph{lavaan} specification of the structural portion of an SEM and
generates all equivalent models according to the "replacing rule" specified by Lee and Hershberger (1990)
}
