% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_trimming.R
\name{equiv_trim}
\alias{equiv_trim}
\title{Trimming Equivalent Models}
\usage{
equiv_trim(equiv_list, prio_vec)
}
\arguments{
\item{equiv_list}{Output of \code{equiv_gen()} or \code{equiv_trim()}, containing a list of equivalent models}

\item{prio_vec}{named numeric vector indicating the priority value for each variable involved in the model.}
}
\value{
A truncated list of equivalent models, in the same format as the initial equiv_list argument supplied.
}
\description{
Given a list of equivalent models (i.e., the output of \code{\link[=equiv_gen]{equiv_gen()}}), alongside a vector specifying the "priority" for each variable
(such that variables with a higher priority cannot cause variables with a lower priority), \code{equiv_trim()} "trims" the number of equivalent models
to retain only those that satisfy the constraints set by the priority vector
}
