% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_boxplot.R
\name{gemini_boxplot}
\alias{gemini_boxplot}
\title{gemini_boxplot}
\usage{
gemini_boxplot(Model, gene.column.names, g, h, nc_gene, sample,
  show_inference = T, color_x = F, identify_guides = F)
}
\arguments{
\item{Model}{a gemini.model object}

\item{gene.column.names}{a character or integer vector of length(2) indicating which columns of \code{guide.annotation} describe the genes being targeted.}

\item{g}{a character naming a gene to visualize}

\item{h}{a character naming another gene to visualize}

\item{nc_gene}{a character naming the gene to use as a negative control.}

\item{sample}{a character naming the sample to visualize}

\item{show_inference}{a logical indicating whether to show the inferred individual or combined values for each gene/gene pair (default T)}

\item{color_x}{a logical indicating whether to visualize the sample-independent effects for each individual guide or guide pair (default F)}

\item{identify_guides}{a logical indicating whether to identify guides with unique colors and shapes (default F)}
}
\value{
a ggplot2 object
}
\description{
A function to visualize the results of GEMINI over the raw data.
}
\details{
Raw LFC data is plotted for each gene combination (`g`-`nc_gene`, `h`-`nc_gene`, `g`-`h`) in a standard boxplot.
Horizontal green line segments are plotted over the box plots indicating the individual gene effects or
the inferred total effect of a particular gene combination. Each guide 
pair can be colored based on the inferred sample independent effects 
\eqn{g_i}, \eqn{h_j}, and \eqn{g_i,h_j}. Additionally, colors and shapes can be used to distinguish unique guides targeting gene g and h, respectively.
}
