% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_conversation.R
\name{gemini_conversation}
\alias{gemini_conversation}
\title{Get Gemini Conversation Endpoint}
\usage{
gemini_conversation(
  prompt,
  gemini_api_key = Sys.getenv("GEMINI_API_KEY"),
  session = NULL
)
}
\arguments{
\item{prompt}{The prompt to start the conversation.}

\item{gemini_api_key}{Gemini's API key.}

\item{session}{Gemini conversation session.}
}
\description{
Get Gemini Conversation Endpoint
}
