% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_image_chat.R
\name{gemini_image_chat}
\alias{gemini_image_chat}
\title{Get Gemini Image Endpoint}
\usage{
gemini_image_chat(prompt, image, gemini_api_key = Sys.getenv("GEMINI_API_KEY"))
}
\arguments{
\item{prompt}{The prompt to start the conversation.}

\item{image}{The path or URL of the image.}

\item{gemini_api_key}{Gemini's API key.}
}
\description{
Get Gemini Image Endpoint
}
