% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_plot_mae.R
\name{gemini_plot_mae}
\alias{gemini_plot_mae}
\title{MAE plot}
\usage{
gemini_plot_mae(Model)
}
\arguments{
\item{Model}{a gemini.model object}
}
\value{
a ggplot2 object
}
\description{
Plots the average MAE of gemini model at each iteration step.
}
\examples{
data("Model", package = "gemini")
gemini_plot_mae(Model)

}
