% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_safety_settings.R
\name{get_safety_settings}
\alias{get_safety_settings}
\title{Get Gemini Safety Settings From Environment Vars}
\usage{
get_safety_settings()
}
\description{
Set environment variables GEMINI_{HARM_CATEGORY_HARASSMENT, HARM_CATEGORY_HATE_SPEECH,
HARM_CATEGORY_SEXUALLY_EXPLICIT, HARM_CATEGORY_DANGEROUS_CONTENT} to affect the result of this
function.
More information about these variables in https://ai.google.dev/gemini-api/docs/safety-settings .
}
