% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{guide.annotation}
\alias{guide.annotation}
\title{Big Papi guide annotations}
\format{An object of class \code{data.frame} with 9216 rows and 7 columns.}
\usage{
guide.annotation
}
\description{
Guide and gene annotations for the Big Papi SynLet 
library.
}
\keyword{data}
