% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_s.R
\name{initialize_s}
\alias{initialize_s}
\title{initialize_s}
\usage{
initialize_s(Model, cores = 1, verbose = FALSE)
}
\arguments{
\item{Model}{an object of class gemini.model}

\item{cores}{a numeric indicating the number of cores to use.  See \code{\link[gemini]{gemini_parallelization}} default 1.}

\item{verbose}{default FALSE}
}
\value{
a Model object of class gemini.model including new slots for s values
}
\description{
Initialize s values using initialized y values and data from \code{Input}
}
\examples{
data("Model", package = "gemini")
Model <- initialize_s(Model)
}
