% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_packages.R
\name{lib.pa}
\alias{lib.pa}
\title{A Lib.pa Function}
\usage{
lib.pa()
}
\description{
This functions allows you to load a bunch of packages with just one line
}
\examples{
lib.pa()
}
\keyword{library}
\keyword{packages}

