% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List Gemini Models}
\usage{
list_models(gemini_api_key = Sys.getenv("GEMINI_API_KEY"))
}
\arguments{
\item{gemini_api_key}{Gemini's API key.}
}
\description{
List Gemini Models
}
