% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{Model}
\alias{Model}
\title{Model object from Big Papi}
\format{
An object of class \code{list} (inherits from \code{gemini.model}) of length 24.
}
\usage{
Model
}
\description{
A gemini.model object created from the Big Papi SynLet dataset after \code{gemini_inference} was run.
The Big Papi dataset was published in Najm et al. 2018 (doi: 10.1038/nbt.4048).
The Model object here is created from the data in \link{Input} using the \code{\link{gemini_initialize}} function.
}
\seealso{
https://www.nature.com/articles/nbt.4048
}
\keyword{data}
