% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{model_numtostring}
\alias{model_numtostring}
\title{Translates model indices to string form (e.g., "AAA")}
\usage{
model_numtostring(model_ind)
}
\arguments{
\item{model_ind}{numeric vector of model indices (e.g., 1, 4, 7)}
}
\description{
Translates model indices to string form (e.g., "AAA")
}
