% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{model_valid}
\alias{model_valid}
\title{Determine model validity}
\usage{
model_valid(permute_df, model_info)
}
\arguments{
\item{permute_df}{permutation dataframe}

\item{model_info}{model object, including information about model index, antecendents, and consequents}
}
\value{
TRUE or FALSE, indicating model validity
}
\description{
Given a permutation dataframe and specific model_info (including antecedent/consequent information), determines if the specified model is "valid"
}
