% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{network_matrix}
\alias{network_matrix}
\title{Network Mapping of Equivalent Models}
\usage{
network_matrix(equiv_list, directed = FALSE, name_labels = FALSE)
}
\arguments{
\item{equiv_list}{Output of \code{equiv_gen()} or \code{equiv_trim()}, containing a list of equivalent models}

\item{directed}{Whether the resultant weights matrix codes unidirectional edges as -1 and bidirectional edges as +1. FALSE by default.}

\item{name_labels}{Whether the column names of the result weights matrix will be each model's model name. FALSE by default, where Model 1 is always the user-specified model. Setting this to TRUE with longer model names may cause readability issues in qgraph.}
}
\value{
matrix representing connectedness of all equivalent models in \code{equiv_list}.
}
\description{
\code{network_matrix} receives the output of \code{\link[=equiv_gen]{equiv_gen()}} or \code{\link[=equiv_trim]{equiv_trim()}} and creates a matrix representing the connectedness between all equivalent models. R package \href{https://www.jstatsoft.org/article/view/v048i04}{qgraph} is required to graph the resultant weights matrix.
}
