% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_trimming.R
\name{param_summary}
\alias{param_summary}
\title{Parameter Summaries of Equivalent Models}
\usage{
param_summary(equiv_list)
}
\arguments{
\item{equiv_list}{Output of \code{equiv_gen()} or \code{equiv_trim()}, containing a list of equivalent models}
}
\value{
dataframe summarizing permutations selected for each parmeter across all equivalent models. Rownames are set as the original permutation specified.
}
\description{
\code{param_summary()} receives the output of \code{\link[=equiv_gen]{equiv_gen()}} or \code{\link[=equiv_trim]{equiv_trim()}} and summarizes the permutations
selected for each parameter across all equivalent models.
}
