% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{pbl_jid_gen}
\alias{pbl_jid_gen}
\title{Generate model names from just-identified preceding block}
\usage{
pbl_jid_gen(pbl_df)
}
\arguments{
\item{pbl_df}{dataframe specifying overall model (i.e., all estimated parameters), with an additional column "pbl_param" indicating if the estimated parameter is "within" the preceding block}
}
\value{
list of model names generated by selecting all possible combinations of permutations involving variables in the preceding block
}
\description{
Given a model specification (dataframe of estimated parameters), generates all models involving permutations to the preceding block
}
