% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{permute_gen}
\alias{permute_gen}
\title{Create permutation dataframe}
\usage{
permute_gen(param_df)
}
\arguments{
\item{param_df}{dataframe of parameters to be estimated, a truncated output of \code{\link[lavaan:lavaanify]{lavaan::lavaanify()}}}
}
\value{
dataframe 3 times as long (# of rows) as the input dataframe, listing all permutations of all parameters
}
\description{
Given a dataframe of free (i.e., not fixed) parameters to be estimated, generates a dataframe containing all permutations (reciprocal relationships excluded) of all parameters
}
