% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{readable_lavaan}
\alias{readable_lavaan}
\title{Create easily readable strings from param df}
\usage{
readable_lavaan(model)
}
\arguments{
\item{model}{single row of a param df (with columns lhs, op, and rhs)}
}
\value{
easily readable character string
}
\description{
given a lavaan row specification (e.g., A ~ B), transforms into a more easily readable string (e.g., B -> A)
}
