% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readg.R
\name{readg}
\alias{readg}
\title{a readg function}
\usage{
readg(site = site, data_name = data_name, dt = FALSE, colClasses = NULL,
  ...)
}
\arguments{
\item{site}{the site of the data to be find, no "" required}

\item{data_name}{the key names of the data required, no "" required}

\item{dt}{whether merged with admit/discharge date/time or not, default false}
}
\description{
This function allows you to read any data files in GEMINI database easily
}
\keyword{gemini}
\keyword{read}

