% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_attributes.R
\name{replace_ind}
\alias{replace_ind}
\title{Replacing model index with given number}
\usage{
replace_ind(model_ind, new_num)
}
\arguments{
\item{model_ind}{numeric vector of model indices (e.g., 1, 4, 7)}

\item{new_num}{Replacement number to be inserted in the correct location}
}
\description{
Replacing model index with given number
}
