% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{replace_rule}
\alias{replace_rule}
\title{Apply replacing rule}
\usage{
replace_rule(permute_df, cur_row, model_info)
}
\arguments{
\item{permute_df}{permutation dataframe}

\item{cur_row}{dataframe of 1 row length specifying focal block under investigation}

\item{model_info}{model info indicating information regarding variable antecedents and consequents}
}
\value{
list of model generated, if they pass the requirements for the replacing rule
}
\description{
Given a permutation dataframe, current focal block under consideration (as a dataframe with a single row), and model info, applies the replacing rule (Lee & Hershberger, 1990)
}
