% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_chat_session.R
\name{reset_chat_session}
\alias{reset_chat_session}
\title{Reset Chat Session}
\usage{
reset_chat_session(session = list())
}
\arguments{
\item{session}{An optional session to set.}
}
\description{
This function is intended to be used with `ask_gemini`. If we are using `ask_gemini` to chat
with Gemini, and we want to start a new conversation, we must call `reset_chat_session`.
}
