% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{sample.replicate.annotation}
\alias{sample.replicate.annotation}
\title{Big Papi sample and replicate annotations}
\format{An object of class \code{data.frame} with 16 rows and 3 columns.}
\usage{
sample.replicate.annotation
}
\description{
Sample and replicate annotations for the Big Papi SynLet dataset.
The Big Papi dataset was published in Najm et al. 2018 (doi: 10.1038/nbt.4048). 
The sample and replicate annotations were acquired from Supplementary Table 3
}
\seealso{
https://www.nature.com/articles/nbt.4048
}
\keyword{data}
