% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Sguide2Pguides_hash}
\alias{Sguide2Pguides_hash}
\title{Guide hashing}
\usage{
Sguide2Pguides_hash(guide2gene, split = ";", cores = 1)
}
\arguments{
\item{guide2gene}{derived from Input object}

\item{split}{character to split guides}

\item{cores}{number of cores to use (default 1)}
}
\value{
Guide hash/list
}
\description{
Guide hashing
}
\examples{
\dontrun{
data("Input", package = "gemini") 
Sguide2Pguides_hash(gemini::Input$guide.pair.annot[1:10,])
}
}
