% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startwith.any.R
\name{startwith.any}
\alias{startwith.any}
\title{A startwith.any Functino}
\usage{
startwith.any(x, prefix)
}
\arguments{
\item{x}{the vector to be compared}

\item{prefix}{the prefixs be be compared with}
}
\description{
This function allows you to see elements in one vector starts with any elements in another vector
}
\examples{
startwith.any(c("aa","bb", "cc", "dd"), c("a", "b"))
}
\keyword{any}
\keyword{startwith,}

