% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{step_gen}
\alias{step_gen}
\title{Single step of generation}
\usage{
step_gen(permute_df, model_info)
}
\arguments{
\item{permute_df}{permutation dataframe}

\item{model_info}{model info indicating information regarding variable antecedents and consequents}
}
\value{
list of all model names of all models potentially equivalent to the original model (does not account for validity)
}
\description{
Given a model (internally generated from a permutation dataframe and model info object), conducts a single "step" of generation where the focal block is iterated across all estimated parameters
}
