% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_to_h.R
\name{swdh}
\alias{swdh}
\title{A Swdh Function}
\usage{
swdh(sitefolder = "")
}
\arguments{
\item{sitefolder}{The site and subfolders in the site}
}
\description{
This functions allows you to set working directory to the local gemini data
}
\examples{
swdr("SMH/CIHI")

}
\keyword{directory}
\keyword{working}

