% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_to_rdrive.R
\name{swdr}
\alias{swdr}
\title{A Swdr Function}
\usage{
swdr(sitefolder = "")
}
\arguments{
\item{sitefolder}{The site and subfolders in the site}
}
\description{
This functions allows you to set working directory to the R drive gemini data
}
\examples{
swdr("SMH/CIHI")

}
\keyword{directory}
\keyword{working}

