% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_s_pb.R
\name{update_s_pb}
\alias{update_s_pb}
\title{update_s}
\usage{
update_s_pb(Model, Input, LFC = "LFC",
  guide.pair.annot = "guide.pair.annot", mean_s = 0, sd_s = 10,
  cores = 1, verbose = F)
}
\arguments{
\item{Model}{a Model object of class gemini.model}

\item{Input}{an Input object of class gemini}

\item{LFC}{an object within \code{Input} containing log-fold change values}

\item{guide.pair.annot}{an object within \code{Input} mapping guide pairs to individual genes.}

\item{mean_s}{numeric indicating prior mean of s (default 0)}

\item{sd_s}{numeric indicating prior sd of s (default 10)}

\item{cores}{a numeric indicating the number of cores to use, see \code{\link{gemini_parallel}}. default=1.}

\item{verbose}{default FALSE}
}
\value{
An object of class gemini.model
}
\description{
Update values of s using data from \code{Input} and current values of other parameters.
}
