% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau_pb.R
\name{update_tau_pb}
\alias{update_tau_pb}
\title{update_tau_pb}
\usage{
update_tau_pb(Model, Input, LFC = "LFC",
  guide.pair.annot = "guide.pair.annot", nc_gene, pattern_join = ";",
  cores = 1, verbose = F)
}
\arguments{
\item{Model}{a Model object of class gemini.model}

\item{Input}{an Input object of class gemini.input}

\item{LFC}{an object within \code{Input} containing log-fold change values}

\item{guide.pair.annot}{an object within \code{Input} mapping guide pairs to individual genes.}

\item{nc_gene}{a character naming the gene to use as a negative control}

\item{pattern_join}{a character to join the gene combinations found in \code{guide.pair.annot}}

\item{cores}{a numeric indicating the number of cores to use.  See \code{\link{gemini_parallel}} for details.  (default=1).}

\item{verbose}{default FALSE}
}
\value{
An object of class gemini.model
}
\description{
Update parameters of tau using data from \code{Input} and current values of other parameters.
}
