% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_y_pb.R
\name{update_y_pb}
\alias{update_y_pb}
\title{update_y_pb}
\usage{
update_y_pb(Model, mean_y = 0, sd_y = 10, verbose = FALSE)
}
\arguments{
\item{Model}{a Model object of class gemini.model}

\item{mean_y}{numeric indicating prior mean of y}

\item{sd_y}{numeric indicating prior sd of y}

\item{verbose}{default FALSE}
}
\value{
An object of class gemini.model
}
\description{
Update values of y using data from \code{Input} and current values of other parameters.
}
