% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{counts}
\alias{counts}
\title{Big Papi counts matrix}
\format{An object of class \code{matrix} with 9216 rows and 16 columns.}
\usage{
counts
}
\description{
Counts matrix from the Big Papi SynLet dataset. 
The Big Papi dataset was published in Najm et al. 2018 (doi: 10.1038/nbt.4048). 
The counts were acquired from Supplementary Table 3.
}
\seealso{
https://www.nature.com/articles/nbt.4048
}
\keyword{data}
