% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_visit_occurrence.R
\name{draw_visit_occurrence}
\alias{draw_visit_occurrence}
\title{Draw visit occurrence}
\usage{
draw_visit_occurrence(std_schema_name, tar_schema_name)
}
\description{
This function for draw graph from visit_occurrence RDS data
}
\keyword{gemini}
