% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini-package.R
\docType{package}
\name{gemini-package}
\alias{gemini}
\alias{gemini-package}
\title{'Google's 'Gemini' <https://gemini.google.com/> coding assistant for 'RStudio'. A set
of functions and 'RStudio' addins that aim to help the R developer in tedious coding tasks.}
\description{
'Google's 'Gemini' \url{https://gemini.google.com/} coding assistant for 'RStudio'. A set of functions and 'RStudio' addins that aim to help the R developer in tedious coding tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcrodriguez1989/gemini}
  \item Report bugs at \url{https://github.com/jcrodriguez1989/gemini/issues}
}

}
\author{
\strong{Maintainer}: Juan Cruz Rodriguez \email{jcrodriguez@unc.edu.ar}

}
