% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{guide.annotation}
\alias{guide.annotation}
\title{Big Papi guide annotations}
\format{An object of class \code{data.frame} with 9216 rows and 7 columns.}
\usage{
guide.annotation
}
\description{
Guide and gene annotations for the Big Papi SynLet dataset. 
The Big Papi dataset was published in Najm et al. 2018 (doi: 10.1038/nbt.4048). 
The guide and gene annotations were acquired from Supplementary Table 2.
}
\seealso{
https://www.nature.com/articles/nbt.4048
}
\keyword{data}
