% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_s.R
\name{initialize_s}
\alias{initialize_s}
\title{initialize_s}
\usage{
initialize_s(Model, Input, LFC, guide.pair.annot = "guide.pair.annot",
  nc_gene, pattern_join = ";", cores = 1, verbose = F)
}
\arguments{
\item{Model}{an object of class gemini.model}

\item{Input}{an object of class gemini.input}

\item{LFC}{an object within \code{Input} containing log-fold change values}

\item{guide.pair.annot}{an object within \code{Input} containing guide to gene annotations}

\item{nc_gene}{a character naming the gene to use as a negative control}

\item{pattern_join}{a character to join the gene combinations found in \code{guide.pair.annot}}

\item{cores}{a numeric indicating the number of cores to use.  See \code{\link[gemini]{gemini_parallel}} default 1.}

\item{verbose}{default FALSE}
}
\value{
a Model object of class gemini.model including new slots for s values
}
\description{
Initialize s values using initialized y values and data from \code{Input}
}
