% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_y.R
\name{initialize_y}
\alias{initialize_y}
\title{initialize_y}
\usage{
initialize_y(Model, verbose = FALSE, cores = 1)
}
\arguments{
\item{Model}{an object of class gemini.model}

\item{verbose}{default FALSE}

\item{cores}{a numeric indicating the number of cores to use.  See details in gemini_parallel. (default=1)}
}
\value{
a Model object of class gemini.model including new slots for y values
}
\description{
Initialize all y values from guide pairs including a negative control.
}
