% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{Input}
\alias{Input}
\alias{Model}
\title{Input object from Big Papi}
\format{An object of class \code{list} (inherits from \code{gemini.input}, \code{gemini.input}) of length 6.}
\usage{
Input

Model
}
\description{
A gemini.input object created from the Big Papi data.

A gemini.model object created from the Big Papi data, after CAVI.
}
\keyword{data}
