% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigpapi-data.R
\docType{data}
\name{sample.replicate.annotation}
\alias{sample.replicate.annotation}
\title{Big Papi sample and replicate annotations}
\format{An object of class \code{data.frame} with 16 rows and 3 columns.}
\usage{
sample.replicate.annotation
}
\description{
Sample and replicate annotations for the Big Papi 
SynLet library.
}
\keyword{data}
