% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mae.R
\name{update_mae}
\alias{update_mae}
\title{update_mae}
\usage{
update_mae(Model, verbose = FALSE)
}
\arguments{
\item{Model}{a Model object of class gemini.model}

\item{verbose}{default FALSE}
}
\value{
An object of class gemini.model
}
\description{
Calculate mean absolute error across all guide combinations.
}
