% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rds_data.R
\name{visit_occurrence_data}
\alias{visit_occurrence_data}
\title{visit occurrence data}
\usage{
visit_occurrence_data(connection)
}
\description{
This function extract data from visit occurrence table
}
\keyword{gemini}
